requirements.txt
tqdm
ipython
matplotlib
moviepy
numpy
Pillow
opencv-python
imutils
scikit-image
scikit-learn
bs4
noise
lxml
dlib
face_recognition
torch
torchvision
tensorflow-gpu
git+https://github.com/tensorpack/tensorpack

